Delete repository suffix
