--- packages/packages/core/editor-interactions-repository/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "@elementor/editor-interactions-repository",
-  "description": "Elementor Editor Interactions Repository",
-  "version": "3.33.0",
-  "private": false,
-  "author": "Elementor Team",
-  "homepage": "https://elementor.com/",
-  "license": "GPL-3.0-or-later",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.mjs",
-      "require": "./dist/index.js"
-    },
-    "./package.json": "./package.json"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/elementor/elementor.git",
-    "directory": "packages/core/editor-interactions-repository"
-  },
-  "bugs": {
-    "url": "https://github.com/elementor/elementor/issues"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "build": "tsup --config=../../tsup.build.ts",
-    "dev": "tsup --config=../../tsup.dev.ts"
-  },
-  "dependencies": {
-    "@elementor/editor-elements": "3.33.0",
-    "@elementor/editor-v1-adapters": "3.33.0",
-    "@elementor/utils": "3.33.0"
-  },
-  "peerDependencies": {
-    "react": "^18.3.1"
-  },
-  "devDependencies": {
-    "tsup": "^8.3.5"
-  }
-}
-
