--- packages/packages/core/editor-interactions-repository/src/index.ts
@@ -1,11 +1,0 @@
-export * from './types';
-
-export { interactionsRepository } from './interactions-repository';
-export {
-	createInteractionsProvider,
-	type CreateInteractionsProviderOptions,
-} from './utils/create-interactions-provider';
-
-export { ELEMENTS_INTERACTIONS_PROVIDER_KEY_PREFIX } from './providers/document-elements-interactions-provider';
-export { init } from './init';
-
