+++ packages/packages/core/editor-interactions/package.json
@@ -42,7 +42,10 @@
         "@elementor/editor-ui": "3.33.0",
         "@elementor/ui": "1.36.15",
         "@elementor/icons": "1.56.0",
-        "@wordpress/i18n": "^5.13.0"
+        "@wordpress/i18n": "^5.13.0",
+        "@elementor/editor-elements": "3.33.0",
+        "@elementor/editor-v1-adapters": "3.33.0",
+        "@elementor/utils": "3.33.0"
     },
     "peerDependencies": {
       "react": "^18.3.1",
