+++ packages/packages/core/editor-interactions/src/errors.ts
