+++ packages/packages/core/editor-interactions/src/index.ts
@@ -3,3 +3,13 @@ export * from './components/interactions-list';
 export * from './types';
 export * from './utils/get-interactions-config';
 export * from './contexts/popup-state-contex';
+
+export { interactionsRepository } from './interactions-repository';
+export {
+	createInteractionsProvider,
+	type CreateInteractionsProviderOptions,
+} from './utils/create-interactions-provider';
+
+export { ELEMENTS_INTERACTIONS_PROVIDER_KEY_PREFIX } from './providers/document-elements-interactions-provider';
+export { init } from './init';
+
