+++ packages/packages/core/editor-interactions/src/types.ts
@@ -21,3 +21,21 @@ export type FieldProps = {
 	onChange: ( value: string ) => void;
 	label?: string;
 };
+
+export type InteractionItem = {
+	elementId: string;
+	dataId: string; // The data-id attribute for DOM selection
+	interactions: string; // JSON string of interactions array
+};
+
+export type InteractionsCollection = InteractionItem[];
+
+export type InteractionsProvider = {
+	getKey: () => string;
+	priority: number;
+	subscribe: ( callback: () => void ) => () => void;
+	actions: {
+		all: () => InteractionItem[];
+	};
+};
+
